# Optimize for parallelism of one