https://github.com/pivovarit/parallel-collectors/issues/417

Before:

```
Benchmark                               (parallelism)   Mode  Cnt      Score     Error  Units
Bench.parallel_batch_collect                        1  thrpt    5  10218.766 ± 131.633  ops/s
Bench.parallel_batch_streaming_collect              1  thrpt    5  10715.432 ±  78.383  ops/s
Bench.parallel_collect                              1  thrpt    5     67.891 ±   1.357  ops/s
Bench.parallel_streaming                            1  thrpt    5     50.920 ±   1.569  ops/s
```

After:
```
Benchmark                               (parallelism)   Mode  Cnt      Score     Error  Units
Bench.parallel_batch_collect                        1  thrpt    5  32793.324 ± 898.335  ops/s
Bench.parallel_batch_streaming_collect              1  thrpt    5  69427.286 ± 448.072  ops/s
Bench.parallel_collect                              1  thrpt    5  32163.622 ± 244.584  ops/s
Bench.parallel_streaming                            1  thrpt    5  69148.860 ± 115.817  ops/s
```

@ Intel i7-4980HQ (8) @ 2.80GHz, 8u222